Put theoretical false positive rate into the mix (cleanup)

+++ src/test/java/test/com/deinersoft/bloomfilter/BloomFilterTest.java
@@ -122,7 +122,7 @@ public class BloomFilterTest {
         System.out.println("actualFalsePositiveRate      = " + String.format("%.2f",(float)(falsePositives)/(float)(actualElementsCount)*100.0)+"%");
 
         // (1 - e^(k * n / m)) ^ k
-        double theoreticalFalsePositiveRate = Math.abs ( Math.pow(1.0 - Math.exp((((double) 1.0 * numberOfHashFunctionsToUse)* (double) actualElementsCount) / (((double) bitsPerElement)*(double)expectedElementsCount)),numberOfHashFunctionsToUse));
+        double theoreticalFalsePositiveRate = Math.abs ( Math.pow(1.0 - Math.exp(((1.0 * numberOfHashFunctionsToUse)* (double) actualElementsCount) / (((double) bitsPerElement)*(double)expectedElementsCount)),numberOfHashFunctionsToUse));
         System.out.println("theoreticalFalsePositiveRate = " + String.format("%.2f",theoreticalFalsePositiveRate*100.0)+"%");
     }
 }
